requirements.txt
langchain==0.3.19
langchain_chroma==0.2.2
langchain_core==0.3.39
langchain_upstage==0.6.0
python-dotenv==1.0.1
python_pptx==1.0.2
streamlit==1.42.2
